Add handlebars, fix bcrypt
Handlebars and the express package to go with it have been added. This,
combined with changes to app.ts, means that we are ready to drop in
handlebars functionality as soon as now. A sample main view has been
added to views/layouts/.

The previous package.json added @types/bcrypt but not bcrypt itself.
This means that the typescript types could be added but bcrypt itself
would not be present during compilation.

Signed-off-by: ahuston-0 <aliceghuston@gmail.com>